requirements.txt
beautifulsoup4==4.13.3
sentence-transformers==3.4.1
Flask==3.1.0
chromadb==0.6.3
pytz==2025.1
uuid==1.30
tzlocal==5.3

